refactor: relocate Container and RichText components to builder directory
