--- src/components/ui/container/index.tsx
@@ -1,9 +1,0 @@
-import { cn } from '@/utils/web/cn'
-
-export function Container(props: { children: React.ReactNode }) {
-  return (
-    <div {...props} className={cn('container')}>
-      {props.children}
-    </div>
-  )
-}
